requirements.txt
Flask>=2.2.2
Pillow
Os